Add rail ASC
Using a base year scenario with rail stopping at Aloha Stadium, I compared the model output of 2500 rail riders to the target of 3500. This is the ridership on the open segment today. I applied an ASC of 0.31 to the rail mode, which is a bonus of roughly 12 minutes of in-vehicle travel time.